# Second Prototype/VisitVietnam-mern/server/docs/apidoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API doc of Places",
    "description": "API doc of Places",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://visitvietnam-api.herokuapp.com",
      "description": "SwaggerHub API VisitVietnam"
    },
    {
      "url": "https://virtserver.swaggerhub.com/challengecup/places/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "http://localhost:5000",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "paths": {
...
    "/reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "get reviews",
        "description": "Reviews\n",
        "operationId": "Reviews",
        "responses": {
          "200": {
            "description": "A list of review log.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/inline_response_200_8"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "inline_response_200_8": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "uuid"
          },
          "StarRating": {
            "type": "number"
          },
          "Pros": {
            "type": "string"
          },
          "Cons": {
            "type": "string"
          },
          "Username": {
            "type": "string"
          },
          "PlaceId": {
            "type": "string"
          },
          "createAt": {
            "type": "string",
            "format": "date"
          },
          "updatedAt": {
            "type": "string",
            "format": "date"
          },
          "__v": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}